backend_admin/pyproject.toml
[tool.poetry]
name = "admin_panel_sprint_2"
version = "0.1.0"
description = ""
authors = ["Michael Fokin <hello@fokin.me>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-extensions = "^3.1.3"
python-dotenv = "^0.19.0"
django-debug-toolbar = "^3.2.1"
gunicorn = "^20.1.0"
django-model-utils = "^4.1.1"
psycopg2-binary = "^2.9.1"
ipython = "^7.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
